# proxy: coalesce identical Series() requests