I spent quite some time recently staring at the fan-out code recently and thought that implementing this would be a good idea: let's join multiple requests for the same data if they are being sent to the same StoreAPI nodes (:

Very often when a new dashboard gets opened, a lot of _different_ queries come in where the same metric data is being used. Clicking around the classic `node_exporter` dashboard shows via the newly implemented metric that this assumption is quite true. It is very unlikely that the retrieved data from SeriesAPI nodes will different if the requests come a few milliseconds between them. So, I think that this is a reasonable optimization to make.

Unfortunately, because upper layers assume sole ownership of the returned `storage.SeriesSet` it means that we have to make copies of the data before passing the data on. There are potential bigger savings here. Added a `TODO` for that.

Finally, I think that caching here is unnecessary because caching of the end-user responses is already performed by query-frontend.

* [x] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

## Changes

Added a `broadcastingSeriesServer` that waits for a response on a query and then fans it out to all of the "listeners".

## Verification

Tests still pass