@GiedriusS 2021-06-07
I have been playing around with groupcache for this and I have some thoughts. I don't think it would work for us in this case at all. This is because the key in groupcache is a simple string and multiple Thanos Query nodes configured in the same group will not always have the same "view" of the leaf nodes. Network errors are inevitable so a key i.e. a series request that will "fall under" some node might not always be doable for that node. So, the keyspace in groupcache would have to be kind of divided into multiple parts where each part has a unique set of StoreAPIs. This is huge work compared to this simple optimization so I'd prefer to go further with the simple variant, and then try to implement the more complex one.