@yeya24 2021-07-11
Looks like this pr works for us and I can see the same `Series` requests are coalesced. This should solve #4407.