@GiedriusS 2021-07-12
Yes, I have tried it. I wouldn't raise a pull request if the code didn't work :smile: The performance question depends on the queries. If there are a lot of them using the same metrics then there is a huge performance win. This still needs a bit of work on the benchmarks. Probably will need to try out https://github.com/lestrrat-go/bufferpool :thinking: