@GiedriusS 2021-07-12
https://github.com/thanos-io/thanos/pull/4379 is also an optimization that implements a similar thing. @yeya24 could you PTAL at it?